Add files via upload
In an attempt to properly study the warfarin dose-response relationship, we have utilized
machine learning to best determine the warfarin dose amount during the initiation period. The
dosage of warfarin is vital to study, as patients are most likely to overdose during this initiation
period and through machine learning we can find the narrow therapeutic index and reduce the
high inter- and intra-individual variability that comes with warfarin. Using data from the
International Warfarin Pharmacogenetics Consortium, we were able to build a regression model
that followed the machine learning model pipeline and ultimately produced results that we were
able to use and evaluate to shed light on warfarin data based on patient data.